+++ packages/tracing/src/browser/index.ts
@@ -1,0 +1,1 @@
+export { BrowserTracing } from './browsertracing';